CreatePost,ImgCloud

+++ Data/Models/Genre.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,7 +8,7 @@ namespace Cirice.Data.Models
 {
     public class Genre
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
         public string GenreString { get; set; }
 
     }
